Add filament density property to basic filament presets
This allows the Plater to calculate and display how much filament will
be used in terms of weight, and also the direct material cost of a print
if the price per spool is specified.

I looked everywhere and couldn't find the density of any FLEX filaments,
so I left that out.